retail-agent/ai_engine/requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
redis
requests
python-dotenv
uvicorn
fastapi
pydantic
streamlit
pandas
requests
reportlab
requests